(PC-36858)[PRO] refactor: rename EventPublicationForm to OfferPublicationForm for more clarity
